moving logic away from index files
